[#11] Feat: Result component 적용
